Added ";", added Bail LAAT, added CrewInfoScreen, set MedicalLAAT as Medical Vehicle, upgraded Pilot Camera for the new ACE
Changes from the 13th of August, 2020:
Added multiple Semikolons in an attempt to let spawned LAATs have the right texture.
Added Bail custom LAAT.
Set MedEvac LAAT to be a Medical Vehicle.
Se tthe flag for the new ACE to check if the Pilot Camera is used to guide wire guided missiles that support the Pilot Camera to do so.